Small corrections. Added sign up js and handlebars. Took sign up function out of login and into it's own js file.
