+++ controllers/homeRoutes.js
@@ -108,4 +108,13 @@ router.get('/login', (req, res) => {
   res.render('login');
 });
 
+router.get('/signup', (req, res) => {
+  if (req.session.logged_in) {
+    res.redirect('/homepage');
+    return;
+  }
+
+  res.render('signup');
+});
+
 module.exports = router;
