+++ models/User.js
@@ -37,7 +37,7 @@ User.init(
         len: [8],
       },
     },
-
+    
     balance: {
       type: DataTypes.INTEGER,
       allowNull: false,
